remove: extra nuget to validate JWT token test:feat: validate JWT token
