+++ QuantConnect.CoinbaseBrokerage/QuantConnect.CoinbaseBrokerage.csproj
@@ -23,16 +23,11 @@
 		<OutputPath>bin\Release\</OutputPath>
 	</PropertyGroup>
 	<ItemGroup>
-		<Compile Include="..\..\Lean\Common\Properties\SharedAssemblyInfo.cs"
-			Link="Properties\SharedAssemblyInfo.cs" />
+		<Compile Include="..\..\Lean\Common\Properties\SharedAssemblyInfo.cs" Link="Properties\SharedAssemblyInfo.cs" />
 	</ItemGroup>
 	<ItemGroup>
 		<PackageReference Include="jose-jwt" Version="5.0.0" PrivateAssets="analyzers;build" />
-		<PackageReference Include="Microsoft.IdentityModel.Tokens" Version="8.0.1"
-			PrivateAssets="analyzers;build" />
 		<PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
 		<PackageReference Include="QuantConnect.Brokerages" Version="2.5.*" />
-		<PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="8.0.1"
-			PrivateAssets="analyzers;build" />
 	</ItemGroup>
 </Project>
